Moved view report router into it's own file
